cloud_function/requirements.txt
functions-framework==3.3.0
pymongo==4.5.0
google-cloud-storage==2.10.0
sentence-transformers==2.2.2
numpy==1.25.2
pandas==2.0.3 